Move env var getters out of `util.ts`
